+++ apps/namadillo/src/App/Common/BalanceChart.tsx
@@ -5,11 +5,12 @@ import {
   SkeletonLoading,
 } from "@namada/components";
 import BigNumber from "bignumber.js";
+import { useState } from "react";
 import { colors } from "theme";
 import { NamCurrency } from "./NamCurrency";
+import { OpacitySlides } from "./OpacitySlides";
 
 type BalanceChartProps = {
-  view: "total" | "stake";
   bondedAmount: BigNumber;
   unbondedAmount: BigNumber;
   availableAmount: BigNumber;
@@ -19,7 +20,6 @@ type BalanceChartProps = {
 };
 
 export const BalanceChart = ({
-  view,
   availableAmount,
   bondedAmount,
   unbondedAmount,
@@ -27,6 +27,8 @@ export const BalanceChart = ({
   totalAmount,
   isLoading,
 }: BalanceChartProps): JSX.Element => {
+  const [activeIndex, setActiveIndex] = useState<number>(0);
+
   const getPiechartData = (): Array<PieChartData> => {
     if (isLoading) {
       return [];
@@ -37,33 +39,25 @@ export const BalanceChart = ({
     }
 
     return [
-      { value: availableAmount, color: colors.balance },
-      { value: bondedAmount, color: colors.bond },
       {
+        label: "Total Staked Balance",
+        value: bondedAmount,
+        color: colors.bond,
+      },
+      {
+        label: "Available NAM to Stake",
+        value: availableAmount,
+        color: colors.balance,
+      },
+      {
+        label: "Unbonding NAM",
         value: unbondedAmount.plus(withdrawableAmount),
         color: colors.unbond,
       },
     ];
   };
 
-  const renderTextSummary = (
-    text: string,
-    balance: BigNumber
-  ): React.ReactNode => {
-    return (
-      <div className="flex flex-col gap-1 leading-tight">
-        <Heading className="text-sm text-neutral-500" level="h3">
-          {text}
-        </Heading>
-        <NamCurrency
-          amount={balance}
-          className="text-2xl"
-          currencySymbolClassName="block mb-1 text-xs ml-1"
-          decimalPlaces={2}
-        />
-      </div>
-    );
-  };
+  const data = getPiechartData();
 
   return (
     <div className="h-[250px] w-[250px]">
@@ -75,14 +69,33 @@ export const BalanceChart = ({
         />
       : <PieChart
           id="balance-chart"
-          data={getPiechartData()}
+          data={data}
           strokeWidth={24}
           radius={125}
           segmentMargin={0}
+          onMouseLeave={() => setActiveIndex(0)}
+          onMouseEnter={(_data: PieChartData, index: number) =>
+            setActiveIndex(index)
+          }
         >
-          {view === "stake" &&
-            renderTextSummary("Total Staked Balance", bondedAmount)}
-          {view === "total" && renderTextSummary("Total Balance", totalAmount)}
+          <OpacitySlides activeIndex={activeIndex}>
+            {data.map((item) => (
+              <div
+                key={item.label}
+                className="flex flex-col gap-1 leading-tight"
+              >
+                <Heading className="text-sm text-neutral-500" level="h3">
+                  {item.label}
+                </Heading>
+                <NamCurrency
+                  amount={item.value}
+                  className="text-2xl"
+                  currencySymbolClassName="block mb-1 text-xs ml-1"
+                  decimalPlaces={2}
+                />
+              </div>
+            ))}
+          </OpacitySlides>
         </PieChart>
       }
     </div>