+++ apps/namadillo/src/App/Settings/Advanced.tsx
@@ -1,14 +1,12 @@
-import { ActionButton, Checkbox, Input, Stack } from "@namada/components";
+import { ActionButton, Input, Stack } from "@namada/components";
 import { chainParametersAtom } from "atoms/chain";
 import {
   indexerUrlAtom,
   maspIndexerUrlAtom,
   rpcUrlAtom,
-  settingsAtom,
   updateIndexerUrlAtom,
   updateMaspIndexerUrlAtom,
   updateRpcUrlAtom,
-  updateSettingsProps,
 } from "atoms/settings";
 import { useAtom, useAtomValue } from "jotai";
 import { useState } from "react";
@@ -16,9 +14,6 @@ import { useLocation } from "react-router-dom";
 
 export const Advanced = (): JSX.Element => {
   const location = useLocation();
-
-  const settings = useAtomValue(settingsAtom);
-  const settingsMutation = useAtomValue(updateSettingsProps);
   const rpcMutation = useAtomValue(updateRpcUrlAtom);
   const indexerMutation = useAtomValue(updateIndexerUrlAtom);
   const [currentMaspIndexer] = useAtom(maspIndexerUrlAtom);
@@ -26,12 +21,8 @@ export const Advanced = (): JSX.Element => {
   const [currentIndexerUrl] = useAtom(indexerUrlAtom);
   const [maspIndexerMutation] = useAtom(updateMaspIndexerUrlAtom);
   const { data: chainParameters } = useAtomValue(chainParametersAtom);
-
   const [rpc, setRpc] = useState(currentRpcUrl);
   const [indexer, setIndexer] = useState(currentIndexerUrl);
-  const [enableAdvancedMode, setAdvancedMode] = useState<boolean>(
-    settings.advancedMode || false
-  );
   const [maspIndexer, setMaspIndexer] = useState(currentMaspIndexer);
 
   const onSubmit = async (e: React.FormEvent): Promise<void> => {
@@ -40,10 +31,6 @@ export const Advanced = (): JSX.Element => {
       await Promise.all([
         rpcMutation.mutateAsync(rpc),
         indexerMutation.mutateAsync(indexer),
-        settingsMutation.mutateAsync({
-          key: "advancedMode",
-          value: enableAdvancedMode,
-        }),
         maspIndexerMutation.mutateAsync(maspIndexer),
       ]);
       document.location.href =
@@ -114,16 +101,6 @@ export const Advanced = (): JSX.Element => {
           disabled={true}
           className="[&_input]:border-neutral-800"
         />
-        <div className="flex gap-3 items-center">
-          <Checkbox
-            id="advanced-mode-enabled"
-            checked={enableAdvancedMode}
-            onChange={(e) => setAdvancedMode(e.target.checked)}
-          />
-          <label htmlFor="advanced-mode-enabled" className="cursor-pointer">
-            Advanced mode
-          </label>
-        </div>
       </Stack>
       <ActionButton
         className="shrink-0"