+++ apps/namadillo/src/App/Staking/StakingRewardsPanel.tsx
@@ -12,13 +12,19 @@ import { sumBigNumberArray } from "utils";
 
 export const StakingRewardsPanel = (): JSX.Element => {
   const { claimRewardsEnabled } = useAtomValue(applicationFeaturesAtom);
-  const { data: rewards } = useAtomValue(claimableRewardsAtom);
+  const {
+    data: rewards,
+    isLoading: rewardsLoading,
+    isSuccess: rewardsLoaded,
+    isRefetching: rewardsRefetching,
+  } = useAtomValue(claimableRewardsAtom);
   const location = useLocation();
   const navigate = useNavigate();
   const availableRewards =
     claimRewardsEnabled ?
       sumBigNumberArray(Object.values(rewards || {}))
     : new BigNumber(0);
+
   const title =
     claimRewardsEnabled ?
       "Unclaimed Staking Rewards"
@@ -29,6 +35,8 @@ export const StakingRewardsPanel = (): JSX.Element => {
       className={clsx({
         "opacity-60 pointer-events-none select-none": !claimRewardsEnabled,
       })}
+      isLoading={rewardsLoading || rewardsRefetching}
+      isSuccess={rewardsLoaded}
       logoElement={
         <i className="text-4xl">
           <GoStack />